Move operators into individual files, use `#[hydroflow_internalmacro::operator_docgen]` macro
