+++ hydroflow_lang/src/graph/ops/flat_map.rs
@@ -1,0 +1,57 @@
+use super::{
+    OperatorConstraints, OperatorWriteOutput, WriteContextArgs, WriteIteratorArgs, RANGE_1,
+};
+
+use quote::quote_spanned;
+
+/// > 1 input stream, 1 output stream
+///
+/// > Arguments: A Rust closure that handles an iterator
+///
+/// For each item `i` passed in, treat `i` as an iterator and map the closure to that
+/// iterator to produce items one by one. the type of the input items must be iterable.
+///
+/// ```hydroflow
+/// recv_iter(vec!["hello", "world"]) -> flat_map(|x| x.chars())
+///     -> for_each(|x| println!("{}", x));
+/// ```
+#[hydroflow_internalmacro::operator_docgen]
+pub const FLAT_MAP: OperatorConstraints = OperatorConstraints {
+    name: "flat_map",
+    hard_range_inn: RANGE_1,
+    soft_range_inn: RANGE_1,
+    hard_range_out: RANGE_1,
+    soft_range_out: RANGE_1,
+    ports_inn: None,
+    ports_out: None,
+    num_args: 1,
+    input_delaytype_fn: &|_| None,
+    write_fn: &(|&WriteContextArgs { root, op_span, .. },
+                 &WriteIteratorArgs {
+                     ident,
+                     inputs,
+                     outputs,
+                     arguments,
+                     is_pull,
+                     ..
+                 }| {
+        let write_iterator = if is_pull {
+            let input = &inputs[0];
+            quote_spanned! {op_span=>
+                let #ident = #input.flat_map(#arguments);
+            }
+        } else {
+            let output = &outputs[0];
+            quote_spanned! {op_span=>
+                let #ident = #root::pusherator::map::Map::new(
+                    #arguments,
+                    #root::pusherator::flatten::Flatten::new(#output)
+                );
+            }
+        };
+        OperatorWriteOutput {
+            write_iterator,
+            ..Default::default()
+        }
+    }),
+};
