+++ hydroflow_lang/src/graph/ops/mod.rs
@@ -1,19 +1,42 @@
 use std::fmt::Display;
 use std::ops::{Bound, RangeBounds};
 
-// Don't use `quote!`, use `quote_spanned!`.
-#[allow(unused_imports)]
-use std::compile_error as quote;
-
 use proc_macro2::{Ident, Span, TokenStream};
-use quote::{quote_spanned, ToTokens};
+use quote::quote_spanned;
 use slotmap::Key;
 use syn::punctuated::Punctuated;
-use syn::{parse_quote, Expr, GenericArgument, Token};
+use syn::{Expr, GenericArgument, Token};
 
 use crate::parse::PortIndex;
 
 use super::{GraphNodeId, GraphSubgraphId, PortIndexValue};
+
+mod cross_join;
+mod difference;
+mod filter;
+mod filter_map;
+mod flat_map;
+mod flatten;
+mod fold;
+mod for_each;
+mod groupby;
+mod identity;
+mod inspect;
+mod join;
+mod map;
+mod merge;
+mod next_epoch;
+mod next_stratum;
+mod null;
+mod recv_iter;
+mod recv_stream;
+mod reduce;
+mod repeat_iter;
+mod sink_async;
+mod sort;
+mod tee;
+mod unique;
+mod write_async;
 
 #[derive(Clone, Copy, PartialOrd, Ord, PartialEq, Eq, Debug)]
 pub enum DelayType {
@@ -107,851 +130,33 @@ pub const IDENTITY_WRITE_FN: &'static dyn Fn(
     }
 });
 
-pub const OPERATORS: [OperatorConstraints; 26] = [
-    OperatorConstraints {
-        name: "null",
-        hard_range_inn: RANGE_ANY,
-        soft_range_inn: RANGE_ANY,
-        hard_range_out: RANGE_ANY,
-        soft_range_out: RANGE_ANY,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|&WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         outputs,
-                         is_pull,
-                         ..
-                     }| {
-            let write_iterator = if is_pull {
-                quote_spanned! {op_span=>
-                    (#(#inputs.for_each(std::mem::drop)),*);
-                    let #ident = std::iter::empty();
-                }
-            } else {
-                quote_spanned! {op_span=>
-                    #[allow(clippy::let_unit_value)]
-                    let _ = (#(#outputs),*);
-                    let #ident = #root::pusherator::for_each::ForEach::new(std::mem::drop);
-                }
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "merge",
-        hard_range_inn: RANGE_ANY,
-        soft_range_inn: &(2..),
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|&WriteContextArgs { op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         outputs,
-                         is_pull,
-                         ..
-                     }| {
-            let write_iterator = if is_pull {
-                let chains = inputs
-                    .iter()
-                    .map(|i| i.to_token_stream())
-                    .reduce(|a, b| quote_spanned! {op_span=> #a.chain(#b) })
-                    .unwrap_or_else(|| quote_spanned! {op_span=> std::iter::empty() });
-                quote_spanned! {op_span=>
-                    let #ident = #chains;
-                }
-            } else {
-                assert_eq!(1, outputs.len());
-                let output = &outputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #output;
-                }
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "join",
-        hard_range_inn: &(2..=2),
-        soft_range_inn: &(2..=2),
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: Some(&(|| parse_quote! { 0, 1 })),
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|wc @ &WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs { ident, inputs, .. }| {
-            let joindata_ident = wc.make_ident("joindata");
-            let write_prologue = quote_spanned! {op_span=>
-                let mut #joindata_ident = Default::default();
-            };
-
-            let lhs = &inputs[0];
-            let rhs = &inputs[1];
-            let write_iterator = quote_spanned! {op_span=>
-                let #ident = #root::compiled::pull::SymmetricHashJoin::new(#lhs, #rhs, &mut #joindata_ident);
-            };
-
-            OperatorWriteOutput {
-                write_prologue,
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "cross_join",
-        hard_range_inn: &(2..=2),
-        soft_range_inn: &(2..=2),
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: Some(&(|| parse_quote! { 0, 1 })),
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|wc @ &WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs { ident, inputs, .. }| {
-            let joindata_ident = wc.make_ident("joindata");
-            let write_prologue = quote_spanned! {op_span=>
-                let mut #joindata_ident = Default::default();
-            };
-
-            let lhs = &inputs[0];
-            let rhs = &inputs[1];
-            let write_iterator = quote_spanned! {op_span=>
-                let #lhs = #lhs.map(|a| ((), a));
-                let #rhs = #rhs.map(|b| ((), b));
-                let #ident = #root::compiled::pull::SymmetricHashJoin::new(#lhs, #rhs, &mut #joindata_ident);
-                let #ident = #ident.map(|((), (a, b))| (a, b));
-            };
-
-            OperatorWriteOutput {
-                write_prologue,
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "tee",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_ANY,
-        soft_range_out: &(2..),
-        ports_inn: None,
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|&WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         outputs,
-                         is_pull,
-                         ..
-                     }| {
-            let write_iterator = if !is_pull {
-                let tees = outputs
-                    .iter()
-                    .rev()
-                    .map(|i| i.to_token_stream())
-                    .reduce(|b, a| quote_spanned! {op_span=> #root::pusherator::tee::Tee::new(#a, #b) })
-                    .unwrap_or_else(
-                        || quote_spanned! {op_span=> #root::pusherator::for_each::ForEach::new(std::mem::drop) },
-                    );
-                quote_spanned! {op_span=>
-                    let #ident = #tees;
-                }
-            } else {
-                assert_eq!(1, inputs.len());
-                let input = &inputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #input;
-                }
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "identity",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|_| None,
-        write_fn: IDENTITY_WRITE_FN,
-    },
-    OperatorConstraints {
-        name: "map",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|&WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         outputs,
-                         arguments,
-                         is_pull,
-                         ..
-                     }| {
-            let write_iterator = if is_pull {
-                let input = &inputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #input.map(#arguments);
-                }
-            } else {
-                let output = &outputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #root::pusherator::map::Map::new(#arguments, #output);
-                }
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "inspect",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|&WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         outputs,
-                         arguments,
-                         is_pull,
-                         ..
-                     }| {
-            let write_iterator = if is_pull {
-                let input = &inputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #input.inspect(#arguments);
-                }
-            } else {
-                let output = &outputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #root::pusherator::inspect::Inspect::new(#arguments, #output);
-                }
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "flat_map",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|&WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         outputs,
-                         arguments,
-                         is_pull,
-                         ..
-                     }| {
-            let write_iterator = if is_pull {
-                let input = &inputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #input.flat_map(#arguments);
-                }
-            } else {
-                let output = &outputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #root::pusherator::map::Map::new(
-                        #arguments,
-                        #root::pusherator::flatten::Flatten::new(#output)
-                    );
-                }
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "flatten",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|&WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         outputs,
-                         is_pull,
-                         ..
-                     }| {
-            let write_iterator = if is_pull {
-                let input = &inputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #input.flatten();
-                }
-            } else {
-                let output = &outputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #root::pusherator::flatten::Flatten::new(#output);
-                }
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "filter_map",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|&WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         outputs,
-                         arguments,
-                         is_pull,
-                         ..
-                     }| {
-            let write_iterator = if is_pull {
-                let input = &inputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #input.filter_map(#arguments);
-                }
-            } else {
-                let output = &outputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #root::pusherator::filter_map::FilterMap::new(#arguments, #output);
-                }
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "filter",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|&WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         outputs,
-                         arguments,
-                         is_pull,
-                         ..
-                     }| {
-            let write_iterator = if is_pull {
-                let input = &inputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #input.filter(#arguments);
-                }
-            } else {
-                let output = &outputs[0];
-                quote_spanned! {op_span=>
-                    let #ident = #root::pusherator::filter::Filter::new(#arguments, #output);
-                }
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "fold",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 2,
-        input_delaytype_fn: &|_| Some(DelayType::Stratum),
-        write_fn: &(|&WriteContextArgs { op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         arguments,
-                         is_pull,
-                         ..
-                     }| {
-            assert!(is_pull);
-            let input = &inputs[0];
-            // TODO(mingwei): Issues if initial value is not copy.
-            // TODO(mingwei): Might introduce the initial value multiple times on scheduling.
-            let write_iterator = quote_spanned! {op_span=>
-                let #ident = std::iter::once(#input.fold(#arguments));
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "reduce",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| Some(DelayType::Stratum),
-        write_fn: &(|&WriteContextArgs { op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         arguments,
-                         is_pull,
-                         ..
-                     }| {
-            assert!(is_pull);
-            let input = &inputs[0];
-            let write_iterator = quote_spanned! {op_span=>
-                let #ident = #input.reduce(#arguments).into_iter();
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "groupby",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 2,
-        input_delaytype_fn: &|_| Some(DelayType::Stratum),
-        write_fn: &(|&WriteContextArgs { op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         arguments,
-                         is_pull,
-                         ..
-                     }| {
-            assert!(is_pull);
-            let input = &inputs[0];
-            let initfn = &arguments[0];
-            let aggfn = &arguments[1];
-            let write_iterator = quote_spanned! {op_span=>
-                let #ident = #input.fold(std::collections::HashMap::new(), |mut ht, nxt| {
-                    let e = ht.entry(nxt.0).or_insert_with(#initfn);
-                    #[allow(clippy::redundant_closure_call)]
-                    (#aggfn)(e, nxt.1);
-                    ht
-                }).into_iter();
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "unique",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|_| Some(DelayType::Stratum),
-        write_fn: &(|&WriteContextArgs { op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         arguments,
-                         is_pull,
-                         ..
-                     }| {
-            assert!(is_pull);
-            let input = &inputs[0];
-            let write_iterator = quote_spanned! {op_span=>
-                let #ident = #input.fold(std::collections::HashSet::new(#arguments), |mut prev, nxt| {prev.insert(nxt); prev}).into_iter();
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "sort",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|_| Some(DelayType::Stratum),
-        write_fn: &(|&WriteContextArgs { op_span, .. },
-                     &WriteIteratorArgs {
-                         ident,
-                         inputs,
-                         arguments,
-                         is_pull,
-                         ..
-                     }| {
-            assert!(is_pull);
-            let input = &inputs[0];
-            let write_iterator = quote_spanned! {op_span=>
-                // TODO(mingwei): unneccesary extra into_iter() then collect()
-                let #ident = #input.collect::<std::collections::BinaryHeap<_>>(#arguments).into_sorted_vec().into_iter();
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "recv_stream",
-        hard_range_inn: RANGE_0,
-        soft_range_inn: RANGE_0,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|wc @ &WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident, arguments, ..
-                     }| {
-            let receiver = &arguments[0];
-            let stream_ident = wc.make_ident("stream");
-            let write_prologue = quote_spanned! {op_span=>
-                let mut #stream_ident = Box::pin(#receiver);
-            };
-            let write_iterator = quote_spanned! {op_span=>
-                let #ident = std::iter::from_fn(|| {
-                    match #root::futures::stream::Stream::poll_next(#stream_ident.as_mut(), &mut std::task::Context::from_waker(&context.waker())) {
-                        std::task::Poll::Ready(maybe) => maybe,
-                        std::task::Poll::Pending => None,
-                    }
-                });
-            };
-            OperatorWriteOutput {
-                write_prologue,
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "recv_iter",
-        hard_range_inn: RANGE_0,
-        soft_range_inn: RANGE_0,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|wc @ &WriteContextArgs { op_span, .. },
-                     &WriteIteratorArgs {
-                         ident, arguments, ..
-                     }| {
-            let iter_ident = wc.make_ident("iter");
-            let write_prologue = quote_spanned! {op_span=>
-                let mut #iter_ident = std::iter::IntoIterator::into_iter(#arguments);
-            };
-            let write_iterator = quote_spanned! {op_span=>
-                let #ident = #iter_ident.by_ref();
-            };
-            OperatorWriteOutput {
-                write_prologue,
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "repeat_iter",
-        hard_range_inn: RANGE_0,
-        soft_range_inn: RANGE_0,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|&WriteContextArgs { op_span, .. },
-                     &WriteIteratorArgs {
-                         ident, arguments, ..
-                     }| {
-            let write_iterator = quote_spanned! {op_span=>
-                let #ident = std::iter::IntoIterator::into_iter(#arguments);
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "difference",
-        hard_range_inn: &(2..=2),
-        soft_range_inn: &(2..=2),
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: Some(&|| parse_quote! { pos, neg }),
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|idx| match idx {
-            PortIndexValue::Path(path) if "neg" == path.to_token_stream().to_string() => {
-                Some(DelayType::Stratum)
-            }
-            _else => None,
-        },
-        write_fn: &(|wc @ &WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs { ident, inputs, .. }| {
-            let handle_ident = wc.make_ident("diffdata_handle");
-            let write_prologue = quote_spanned! {op_span=>
-                let #handle_ident = df.add_state(std::cell::RefCell::new(
-                    #root::lang::monotonic_map::MonotonicMap::<_, std::collections::HashSet<_>>::default(),
-                ));
-            };
-            let write_iterator = {
-                let borrow_ident = wc.make_ident("borrow");
-                let negset_ident = wc.make_ident("negset");
-
-                let input_neg = &inputs[0]; // N before P
-                let input_pos = &inputs[1];
-                quote_spanned! {op_span=>
-                    let mut #borrow_ident = context.state_ref(#handle_ident).borrow_mut();
-                    let #negset_ident = #borrow_ident
-                        .try_insert_with((context.current_epoch(), context.current_stratum()), || {
-                            #input_neg.collect()
-                        });
-                    let #ident = #input_pos.filter(move |x| !#negset_ident.contains(x));
-                }
-            };
-            OperatorWriteOutput {
-                write_prologue,
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "next_stratum",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|_| Some(DelayType::Stratum),
-        write_fn: IDENTITY_WRITE_FN,
-    },
-    OperatorConstraints {
-        name: "next_epoch",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_1,
-        soft_range_out: RANGE_1,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 0,
-        input_delaytype_fn: &|_| Some(DelayType::Epoch),
-        write_fn: IDENTITY_WRITE_FN,
-    },
-    OperatorConstraints {
-        name: "for_each",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_0,
-        soft_range_out: RANGE_0,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|&WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident, arguments, ..
-                     }| {
-            let write_iterator = quote_spanned! {op_span=>
-                let #ident = #root::pusherator::for_each::ForEach::new(#arguments);
-            };
-            OperatorWriteOutput {
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "write_async",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_0,
-        soft_range_out: RANGE_0,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|wc @ &WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident, arguments, ..
-                     }| {
-            let async_write_arg = &arguments[0];
-
-            let send_ident = wc.make_ident("item_send");
-            let recv_ident = wc.make_ident("item_recv");
-
-            let write_prologue = quote_spanned! {op_span=>
-                let (#send_ident, #recv_ident) = #root::tokio::sync::mpsc::unbounded_channel();
-                df
-                    .spawn_task(async move {
-                        use #root::tokio::io::AsyncWriteExt;
-
-                        let mut recv = #recv_ident;
-                        let mut write = #async_write_arg;
-                        while let Some(item) = recv.recv().await {
-                            let bytes = std::convert::AsRef::<[u8]>::as_ref(&item);
-                            write.write_all(bytes).await.expect("Error processing async write item.");
-                        }
-                    })
-                    .expect("write_async() must be used within a tokio runtime");
-            };
-
-            let write_iterator = quote_spanned! {op_span=>
-                let #ident = #root::pusherator::for_each::ForEach::new(|item| {
-                    #send_ident.send(item).expect("Failed to send async write item for processing.");
-                });
-            };
-
-            OperatorWriteOutput {
-                write_prologue,
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
-    OperatorConstraints {
-        name: "sink_async",
-        hard_range_inn: RANGE_1,
-        soft_range_inn: RANGE_1,
-        hard_range_out: RANGE_0,
-        soft_range_out: RANGE_0,
-        ports_inn: None,
-        ports_out: None,
-        num_args: 1,
-        input_delaytype_fn: &|_| None,
-        write_fn: &(|wc @ &WriteContextArgs { root, op_span, .. },
-                     &WriteIteratorArgs {
-                         ident, arguments, ..
-                     }| {
-            let sink_arg = &arguments[0];
-
-            let send_ident = wc.make_ident("item_send");
-            let recv_ident = wc.make_ident("item_recv");
-
-            let write_prologue = quote_spanned! {op_span=>
-                let (#send_ident, #recv_ident) = #root::tokio::sync::mpsc::unbounded_channel();
-                df
-                    .spawn_task(async move {
-                        use #root::futures::sink::SinkExt;
-
-                        let mut recv = #recv_ident;
-                        let mut sink = #sink_arg;
-                        while let Some(item) = recv.recv().await {
-                            sink.feed(item).await.expect("Error processing async sink item.");
-                            // Receive as many items synchronously as possible before flushing.
-                            while let Ok(item) = recv.try_recv() {
-                                sink.feed(item).await.expect("Error processing async sink item.");
-                            }
-                            sink.flush().await.expect("Failed to flush async sink.");
-                        }
-                    })
-                    .expect("sink_async() must be used within a tokio runtime");
-            };
-
-            let write_iterator = quote_spanned! {op_span=>
-                let #ident = #root::pusherator::for_each::ForEach::new(|item| {
-                    #send_ident.send(item).expect("Failed to send async write item for processing.");
-                });
-            };
-
-            OperatorWriteOutput {
-                write_prologue,
-                write_iterator,
-                ..Default::default()
-            }
-        }),
-    },
+pub const OPERATORS: &[OperatorConstraints] = &[
+    null::NULL,
+    merge::MERGE,
+    join::JOIN,
+    cross_join::CROSS_JOIN,
+    tee::TEE,
+    identity::IDENTITY,
+    map::MAP,
+    inspect::INSPECT,
+    flat_map::FLAT_MAP,
+    flatten::FLATTEN,
+    filter_map::FILTER_MAP,
+    filter::FILTER,
+    fold::FOLD,
+    reduce::REDUCE,
+    groupby::GROUPBY,
+    unique::UNIQUE,
+    sort::SORT,
+    recv_stream::RECV_STREAM,
+    recv_iter::RECV_ITER,
+    repeat_iter::REPEAT_ITER,
+    difference::DIFFERENCE,
+    next_stratum::NEXT_STRATUM,
+    next_epoch::NEXT_EPOCH,
+    for_each::FOR_EACH,
+    write_async::WRITE_ASYNC,
+    sink_async::SINK_ASYNC,
 ];
 
 pub struct WriteContextArgs<'a> {
